@@ -4,10 +4,10 @@ object CamelCaseFormatter : BaseCaseFormatter() {
     override fun formatWord(index: Int, word: CharSequence): CharSequence {
         return buildString {
             if (index == 0) {
-                return word.toString().toLowerCase()
+                return word.toString().lowercase()
             } else {
-                append(word.first().toUpperCase())
-                append(word.substring(1).toLowerCase())
+                append(word.first().uppercase())
+                append(word.substring(1).lowercase())
             }
         }
     }
