@@ -2,7 +2,7 @@ package io.kanro.idea.plugin.protobuf.string.case
 
 object KebabCaseFormatter : BaseCaseFormatter() {
     override fun formatWord(index: Int, word: CharSequence): CharSequence {
-        return word.toString().toLowerCase()
+        return word.toString().lowercase()
     }
 
     override fun appendDelimiter(builder: StringBuilder) {
