@@ -3,8 +3,8 @@ package io.kanro.idea.plugin.protobuf.string.case
 object PascalCaseFormatter : BaseCaseFormatter() {
     override fun formatWord(index: Int, word: CharSequence): CharSequence {
         return buildString {
-            append(word.first().toUpperCase())
-            append(word.substring(1).toLowerCase())
+            append(word.first().uppercase())
+            append(word.substring(1).lowercase())
         }
     }
 }
